+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Changed
     - Rename ``revoke_failed_invite()`` to ``revoke_invite()``
     - Do not return ``submit_status`` from project API views (#971)
     - Remove required ``owner`` argument for ``ProjectUpdateAPIView`` (#1007)
+    - Remove unused owner operations from ``ProjectModifyMixin`` (#1008)
 
 Fixed
 -----